Updated application review
